// src/types/types.ts
export interface DecodedToken {
  user: User;
  exp: number;
  // Add other JWT claims as needed
  iat?: number;
  iss?: string;
}
  
  export interface User {
  firstName: string;
  email: string;
  siteId?: string;
}